day7/requirements.txt
langchain==0.3.0
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
faiss-cpu
python-dotenv